feat: add input text
