+++ src/utils/fnKeys.js
@@ -1,0 +1,17 @@
+const fnKey = {
+  Tab: false,
+  CapsLock: false,
+  Shift: false,
+  Enter: false,
+  Ctrl: true,
+  Win: true,
+  Alt: true,
+  Space: false,
+  AltRight: true,
+  ArrowLeft: true,
+  ArrowDown: true,
+  ArrowRight: true,
+  CtrlRight: true,
+};
+
+export default fnKey;
