+++ src/utils/keysLayout.js
@@ -1,0 +1,78 @@
+const keysLayout = [
+  [
+    'Tilda',
+    '1',
+    '2',
+    '3',
+    '4',
+    '5',
+    '6',
+    '7',
+    '8',
+    '9',
+    '0',
+    '-',
+    '+',
+    'Backspace',
+  ],
+  [
+    'Tab',
+    'q',
+    'w',
+    'e',
+    'r',
+    't',
+    'y',
+    'u',
+    'i',
+    'o',
+    'p',
+    '[',
+    ']',
+    'BackSlash',
+    'Del',
+  ],
+  [
+    'CapsLock',
+    'a',
+    's',
+    'd',
+    'f',
+    'g',
+    'h',
+    'j',
+    'k',
+    'l',
+    'Semicolon',
+    'Quote',
+    'Enter',
+  ],
+  [
+    'Shift',
+    'z',
+    'x',
+    'c',
+    'v',
+    'b',
+    'n',
+    'm',
+    ',',
+    '.',
+    '/',
+    'ArrowUp',
+    'ShiftRight',
+  ],
+  [
+    'Ctrl',
+    'Win',
+    'Alt',
+    'Space',
+    'AltRight',
+    'ArrowLeft',
+    'ArrowDown',
+    'ArrowRight',
+    'CtrlRight',
+  ],
+];
+
+export default keysLayout;
